updated some code styling for the event views
