--- Olympsis/Views/Viewers/Event/EventFieldView.swift
@@ -1,9 +1,0 @@
-//
-//  EventDetailMapView.swift
-//  Olympsis
-//
-//  Created by Joel Joseph on 1/13/23.
-//
-
-import SwiftUI
-
