--- Olympsis/Views/Viewers/Event/EventHostClubView.swift
@@ -1,9 +1,0 @@
-//
-//  EventHostClubView.swift
-//  Olympsis
-//
-//  Created by Joel Joseph on 1/13/23.
-//
-
-import SwiftUI
-
